+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Don't write Assignment Task GUID or Assignment Resource GUID to JSON.
 * Don't write a value for Assignment Work Contour to JSON if the contour is the default value (i.e. flat).
 * Don't write a value for Assignment Resource Request Type to JSON if the type is the default value (i.e. none).
+* Don't write a value for Task Earned Value Method to JSON if the method matches the project default.
+* Don't write a value for Task Type to JSON if the type matches the project default.
 
 ## 9.4.0 (11/6/2021)
 * Read custom value lists for resource custom fields from MPP files (based on a suggestion by Markus Höger).